Minor tweaks for debugging
Removed print statements as they were filling up the syslog as it runs
as a service.
Removed PiGMi path for files and left it generic.
Fixed some oddities with new Jesse release. This version was installed
on a brand new SD card with a brand new OS. Currently running on a Pi3